+++ pkg/aws/awscontrolplane/mutate_awscontrolplane_preHA_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/aws-admission-controller/pkg/admission"
+	"github.com/giantswarm/aws-admission-controller/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/pkg/unittest"
 )
 
@@ -79,7 +79,7 @@ func TestPreHAAWSControlPlaneAdmit(t *testing.T) {
 				}
 			}
 			fakeK8sClient := unittest.FakeK8sClient()
-			admit := &Admitter{
+			mutate := &Mutator{
 				validAvailabilityZones: []string{"cn-south-1a"},
 				k8sClient:              fakeK8sClient,
 				logger:                 newLogger,
@@ -91,12 +91,12 @@ func TestPreHAAWSControlPlaneAdmit(t *testing.T) {
 				t.Fatal(err)
 			}
 			// run admission request to default AWSControlPlane
-			var patch []admission.PatchOperation
+			var patch []mutator.PatchOperation
 			request, err := awsControlPlaneAdmissionRequest(tc.currentAvailabilityZone, tc.currentInstanceType, preHAReleaseVersion)
 			if err != nil {
 				t.Fatal(err)
 			}
-			patch, err = admit.Admit(request)
+			patch, err = mutate.Mutate(request)
 			if err != nil {
 				t.Fatal(err)
 			}
