+++ pkg/aws/g8scontrolplane/mutate_g8scontrolplane_infraref_test.go
@@ -14,7 +14,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/giantswarm/aws-admission-controller/pkg/admission"
+	"github.com/giantswarm/aws-admission-controller/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/pkg/unittest"
 )
 
@@ -58,7 +58,7 @@ func TestInfraRefG8sControlPlaneAdmit(t *testing.T) {
 				}
 			}
 			fakeK8sClient := unittest.FakeK8sClient()
-			admit := &Admitter{
+			mutate := &Mutator{
 				validAvailabilityZones: []string{"eu-central-1a", "eu-central-1b", "eu-central-1c"},
 				k8sClient:              fakeK8sClient,
 				logger:                 newLogger,
@@ -72,12 +72,12 @@ func TestInfraRefG8sControlPlaneAdmit(t *testing.T) {
 			}
 
 			// run admission request for g8sControlPlane without reference
-			var patch []admission.PatchOperation
+			var patch []mutator.PatchOperation
 			request, err := g8sControlPlaneNoReferenceAdmissionRequest()
 			if err != nil {
 				t.Fatal(err)
 			}
-			patch, err = admit.Admit(request)
+			patch, err = mutate.Mutate(request)
 			if err != nil {
 				t.Fatal(err)
 			}
